Fix manyFinally'
Ignore-this: 4ba68b36c254c355ac303e0b0b400be

Would previously not work correctly on cases like:

> runParser' (manyFinally' (satisfy (<10)) (satisfy (==10))) [1,2,20]

In that the initial parser succeeds but the finally parser doesn't.

This is because the oneOf' call will fail (as
required/expected)... which makes the recursion fail, and thus it
backtracks up a level and on to the oneOf' call in the previous level,
which succeeds (as there's still an initial parser that can succeed).

The commit usage prevents such backtracking, and the explicit use of
<|> means we get the expected error messages rather than the default
one from oneOf.

darcs-hash:20140111083805-9c7ab-59797e6663e1212e3b4b7f89a83cffebfa0876bd